refactor: Move commands into separate file
This in preparation of adding a new "Measure Coverage" command, but I
don't want to make the main `extension.ts` file even longer.
