+++ src/extension/extension.ts
@@ -19,14 +19,7 @@ import {
   ServerOptions,
 } from "vscode-languageclient/node";
 
-import {
-  BazelWorkspaceInfo,
-  activateTaskProvider,
-  createBazelTask,
-  queryQuickPickPackage,
-  queryQuickPickTargets,
-} from "../bazel";
-import { IBazelCommandAdapter } from "../bazel";
+import { activateTaskProvider } from "../bazel";
 import {
   BuildifierDiagnosticsManager,
   BuildifierFormatProvider,
@@ -37,9 +30,9 @@ import { BazelCompletionItemProvider } from "../completion-provider";
 import { BazelGotoDefinitionProvider } from "../definition/bazel_goto_definition_provider";
 import { BazelTargetSymbolProvider } from "../symbols";
 import { BazelWorkspaceTreeProvider } from "../workspace-tree";
-import { getDefaultBazelExecutablePath } from "./configuration";
 import { activateCommandVariables } from "./command_variables";
 import { activateTesting } from "../test-explorer";
+import { activateCommands } from "./commands";
 
 /**
  * Called when the extension is activated; that is, when its first command is
@@ -99,24 +92,7 @@ export async function activate(context: vscode.ExtensionContext) {
       workspaceTreeProvider,
     ),
     // Commands
-    vscode.commands.registerCommand("bazel.buildTarget", bazelBuildTarget),
-    vscode.commands.registerCommand(
-      "bazel.buildTargetWithDebugging",
-      bazelBuildTargetWithDebugging,
-    ),
-    vscode.commands.registerCommand("bazel.buildAll", bazelbuildAll),
-    vscode.commands.registerCommand(
-      "bazel.buildAllRecursive",
-      bazelbuildAllRecursive,
-    ),
-    vscode.commands.registerCommand("bazel.runTarget", bazelRunTarget),
-    vscode.commands.registerCommand("bazel.testTarget", bazelTestTarget),
-    vscode.commands.registerCommand("bazel.testAll", bazelTestAll),
-    vscode.commands.registerCommand(
-      "bazel.testAllRecursive",
-      bazelTestAllRecursive,
-    ),
-    vscode.commands.registerCommand("bazel.clean", bazelClean),
+    ...activateCommands(),
     vscode.commands.registerCommand("bazel.refreshBazelBuildTargets", () => {
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
       completionItemProvider.refresh();
@@ -183,295 +159,4 @@ function createLsp(config: vscode.WorkspaceConfiguration) {
   };
 
   return new LanguageClient("Bazel LSP Client", serverOptions, clientOptions);
-}
-
-/**
- * Builds a Bazel target and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelBuildTarget(adapter: IBazelCommandAdapter | undefined) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick build targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickTargets({ query: "kind('.* rule', ...)" }),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      await bazelBuildTarget(quickPick);
-    }
-    return;
-  }
-  const commandOptions = adapter.getBazelCommandOptions();
-  const task = createBazelTask("build", commandOptions);
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Builds a Bazel target and attaches the Starlark debugger.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelBuildTargetWithDebugging(
-  adapter: IBazelCommandAdapter | undefined,
-) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick build targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickTargets({ query: "kind('.* rule', ...)" }),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      bazelBuildTargetWithDebugging(quickPick);
-    }
-    return;
-  }
-  const bazelConfigCmdLine =
-    vscode.workspace.getConfiguration("bazel.commandLine");
-  const startupOptions = bazelConfigCmdLine.get<string[]>("startupOptions");
-  const commandArgs = bazelConfigCmdLine.get<string[]>("commandArgs");
-
-  const commandOptions = adapter.getBazelCommandOptions();
-
-  const fullArgs = commandArgs
-    .concat(commandOptions.targets)
-    .concat(commandOptions.options);
-
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.debug.startDebugging(undefined, {
-    args: fullArgs,
-    bazelCommand: "build",
-    bazelExecutablePath: getDefaultBazelExecutablePath(),
-    bazelStartupOptions: startupOptions,
-    cwd: commandOptions.workspaceInfo.bazelWorkspacePath,
-    name: "On-demand Bazel Build Debug",
-    request: "launch",
-    type: "bazel-launch-build",
-  });
-}
-
-/**
- * Builds a Bazel package and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelbuildAll(adapter: IBazelCommandAdapter | undefined) {
-  await buildPackage(":all", adapter);
-}
-
-/**
- * Builds a Bazel package recursively and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelbuildAllRecursive(
-  adapter: IBazelCommandAdapter | undefined,
-) {
-  await buildPackage("/...", adapter);
-}
-
-async function buildPackage(
-  suffix: string,
-  adapter: IBazelCommandAdapter | undefined,
-) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick build targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickPackage({}),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      await buildPackage(suffix, quickPick);
-    }
-    return;
-  }
-  const commandOptions = adapter.getBazelCommandOptions();
-  const allCommandOptions = {
-    options: commandOptions.options,
-    targets: commandOptions.targets.map((s) => s + suffix),
-    workspaceInfo: commandOptions.workspaceInfo,
-  };
-  const task = createBazelTask("build", allCommandOptions);
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Runs a Bazel target and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelRunTarget(adapter: IBazelCommandAdapter | undefined) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick test targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickTargets({ query: "kind('.* rule', ...)" }),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      await bazelRunTarget(quickPick);
-    }
-    return;
-  }
-  const commandOptions = adapter.getBazelCommandOptions();
-  const task = createBazelTask("run", commandOptions);
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Tests a Bazel target and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelTestTarget(adapter: IBazelCommandAdapter | undefined) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick test targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickTargets({ query: "kind('.*_test rule', ...)" }),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      await bazelTestTarget(quickPick);
-    }
-    return;
-  }
-  const commandOptions = adapter.getBazelCommandOptions();
-  const task = createBazelTask("test", commandOptions);
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Tests a Bazel package and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelTestAll(adapter: IBazelCommandAdapter | undefined) {
-  await testPackage(":all", adapter);
-}
-
-/**
- * Tests a Bazel package recursively and streams output to the terminal.
- *
- * @param adapter An object that implements {@link IBazelCommandAdapter} from
- * which the command's arguments will be determined.
- */
-async function bazelTestAllRecursive(
-  adapter: IBazelCommandAdapter | undefined,
-) {
-  await testPackage("/...", adapter);
-}
-
-async function testPackage(
-  suffix: string,
-  adapter: IBazelCommandAdapter | undefined,
-) {
-  if (adapter === undefined) {
-    // If the command adapter was unspecified, it means this command is being
-    // invoked via the command palatte. Provide quickpick build targets for
-    // the user to choose from.
-    const quickPick = await vscode.window.showQuickPick(
-      queryQuickPickPackage({}),
-      {
-        canPickMany: false,
-      },
-    );
-    // If the result was undefined, the user cancelled the quick pick, so don't
-    // try again.
-    if (quickPick) {
-      await testPackage(suffix, quickPick);
-    }
-    return;
-  }
-  const commandOptions = adapter.getBazelCommandOptions();
-  const allCommandOptions = {
-    options: commandOptions.options,
-    targets: commandOptions.targets.map((s) => s + suffix),
-    workspaceInfo: commandOptions.workspaceInfo,
-  };
-  const task = createBazelTask("test", allCommandOptions);
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Cleans a Bazel workspace.
- *
- * If there is only a single workspace open, it will be cleaned immediately. If
- * there are multiple workspace folders open, a quick-pick window will be opened
- * asking the user to choose one.
- */
-async function bazelClean() {
-  const workspaceInfo = await BazelWorkspaceInfo.fromWorkspaceFolders();
-  if (!workspaceInfo) {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    vscode.window.showInformationMessage(
-      "Please open a Bazel workspace folder to use this command.",
-    );
-
-    return;
-  }
-  const task = createBazelTask("clean", {
-    options: [],
-    targets: [],
-    workspaceInfo,
-  });
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.tasks.executeTask(task);
-}
-
-/**
- * Copies a target to the clipboard.
- */
-function bazelCopyTargetToClipboard(adapter: IBazelCommandAdapter | undefined) {
-  if (adapter === undefined) {
-    // This command should not be enabled in the commands palette, so adapter
-    // should always be present.
-    return;
-  }
-  // This can only be called on single targets, so we can assume there is only
-  // one of them.
-  const target = adapter.getBazelCommandOptions().targets[0];
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  vscode.env.clipboard.writeText(target);
 }
